Fixed a bug with the playgame cpp and header files, added gameloop function to playgame and called it on the main file, also the Initialize, Close and GameLoop function were defined

+++ Src/Objects/Frog.cpp
@@ -15,5 +15,5 @@ Frog CreateFrog(Frog& frog)
 
 void DrawFrog(Frog frog)
 {
-	DrawRectangle(frog.frogPosition.x, frog.frogPosition.y, frog.frogSize.x, frog.frogSize.y, frog.frogColor);
+	DrawRectangle(static_cast<int>(frog.frogPosition.x), static_cast<int>(frog.frogPosition.y), static_cast<int>(frog.frogSize.x), static_cast<int>(frog.frogSize.y), frog.frogColor);
 }
